+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fixes starfield polarization; checks times are in UTC in https://github.com/punch-mission/punchbowl/pull/328#pullrequestreview-2726230483
 - Fixes issues with calibration metadata in https://github.com/punch-mission/punchbowl/pull/404
 - Fixes a keyword typo in omniheader in https://github.com/punch-mission/punchbowl/pull/407
+- Creates a uses a custom PUNCH flow in prefect in https://github.com/punch-mission/punchbowl/pull/409
 - Changes default of NDCube writing to prohibit overwriting in https://github.com/punch-mission/punchbowl/pull/408
 
 ## Version 0.0.9: Feb 28, 2025